requirements.txt
numpy
pillow
soundfile
torch
torchvision
matplotlib
tqdm
resampy
wandb
nnAudio
librosa
celluloid
opencv-python
